Fix Friendship model to always return your friend
The Friendship model assumed that the first user listed in the friendship
was your friend. However, in some cases the first user listed is yourself.
To correct this issue, I added a 'Self' property to the friendship, which
is populated by the proxy. When the 'Friend' property is queried we return
the user that *isn't* yourself.